requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
bcrypt==4.1.3
slowapi==0.1.9
aiosmtplib==3.0.1
python-dotenv==1.0.1